♻️: Refactoring before handing in the project
