+++ frontend/app_student/lib/shared_components/app_bar.dart
