+++ frontend/app_student/lib/shared_components/header_logo.dart
